Spliff off all individual OverviewScreen panes
